+++ demo/run.sh
@@ -13,24 +13,24 @@ if ! type python > /dev/null 2>&1 && ! type python3 > /dev/null 2>&1; then
     exit 1
 fi
 
-if ! type poetry > /dev/null 2>&1; then
-    echo "Cannot start the canaille demo server. Please install poetry on your system"
+if ! type uv > /dev/null 2>&1; then
+    echo "Cannot start the canaille demo server. Please install uv on your system"
     echo "or run the demo with docker-compose."
-    echo "https://python-poetry.org/docs/#installation"
+    echo "https://docs.astral.sh/uv/getting-started/installation/"
     exit 1
 fi
 
 pushd "$DIR" > /dev/null 2>&1 || exit
 
-poetry install --with demo --all-extras
+uv sync --group demo --all-extras
 
 if [ "$BACKEND" = "memory" ]; then
 
-    poetry run honcho --env ../.env --procfile Procfile-memory start
+    uv run honcho --env ../.env --procfile Procfile-memory start
 
 elif [ "$BACKEND" = "sql" ]; then
 
-    poetry run honcho --env ../.env --procfile Procfile-sql start
+    uv run honcho --env ../.env --procfile Procfile-sql start
 
 elif [ "$BACKEND" = "ldap" ]; then
 
@@ -40,7 +40,7 @@ elif [ "$BACKEND" = "ldap" ]; then
         exit 1
     fi
 
-    poetry run honcho --env ../.env --procfile Procfile-ldap start
+    uv run honcho --env ../.env --procfile Procfile-ldap start
 
 else